Add files via upload
Part 1- Dataset for Sales and Resturant. By the dataset given we can find the performance of an organization through their sales data. We are getting the information regarding the sales of each products, delivery status and performance, profit or loss of each product, customers who are regular with the orders and placed maximum orders.
Part 2- Dataset for Resturant. By the dataset given we can find out the number of cuisine on demand in different geolocations, ratings from the customers for different cuisines and service, also the parking availability under each restaurant and payment methods offer by restaurant and the customers.